+++ src/main/java/dev/tazer/mixed_litter/client/renderers/SquidRemodelRenderer.java
@@ -2,9 +2,8 @@ package dev.tazer.mixed_litter.client.renderers;
 
 import dev.tazer.mixed_litter.client.ModelLayers;
 import dev.tazer.mixed_litter.client.models.SquidRemodel;
-import net.minecraft.client.model.SquidModel;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
-import net.minecraft.client.renderer.entity.MobRenderer;
+import net.minecraft.client.renderer.entity.SquidRenderer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.GlowSquid;
@@ -12,11 +11,11 @@ import net.minecraft.world.entity.animal.Squid;
 
 import java.util.function.Function;
 
-public class SquidRemodelRenderer extends MobRenderer<Squid, SquidModel<Squid>> implements RemodelMarker {
+public class SquidRemodelRenderer extends SquidRenderer<Squid> implements RemodelMarker {
     private final Function<Squid, ResourceLocation> texture;
 
     public SquidRemodelRenderer(EntityRendererProvider.Context context, Function<Squid, ResourceLocation> texture) {
-        super(context, new SquidRemodel<>(context.bakeLayer(ModelLayers.SQUID_LAYER)), 0.7F);
+        super(context, new SquidRemodel<>(context.bakeLayer(ModelLayers.SQUID_LAYER)));
         this.texture = texture;
     }